비급여 합계 int -> long 데이터 형 수정

+++ src/main/java/com/eis_project/revenue/api/reponse/RevenueDeptResponse.java
@@ -22,5 +22,5 @@ import lombok.Setter;
 public class RevenueDeptResponse {
     private String deptCd;
     private String deptNm;
-    private Integer totAmt;
+    private Long totAmt;
 }
